Restructure project to seperate ContractABI subspec/SPM module
Explicitely depending on `Web3/PromiseKit` will also import the Promise
extensions for the ContractABI module if and only if the ContractABI
subspec is used.
