+++ Example/Podfile
@@ -2,6 +2,7 @@ use_frameworks!
 target 'Web3_Tests' do
   pod 'Web3', :path => '../'
   pod 'Web3/PromiseKit', :path => '../'
+  pod 'Web3/ContractABI', :path => '../'
   
   pod 'Quick', '~> 1.2.0'
   pod 'Nimble', '~> 7.0.2'
