+++ Example/Tests/ABITests/SolidityTypeTests.swift
@@ -9,6 +9,9 @@
 import XCTest
 import Web3
 import BigInt
+#if canImport(Web3ContractABI)
+    @testable import Web3ContractABI
+#endif
 
 class SolidityTypeTests: XCTestCase {
