+++ Web3/Classes/ContractABI/ABI/ABI.swift
@@ -7,6 +7,9 @@
 
 import Foundation
 import BigInt
+#if !Web3CocoaPods
+    import Web3
+#endif
 
 /// ABI Namespace
 public struct ABI {
