+++ Web3/Classes/ContractABI/ABI/Eth+ABI.swift
@@ -6,6 +6,9 @@
 //
 
 import Foundation
+#if !Web3CocoaPods
+    import Web3
+#endif
 
 public extension Web3.Eth {
