+++ Web3/Classes/ContractABI/Contract/ContractPromiseExtensions.swift
@@ -1,0 +1,47 @@
+//
+//  ContractPromiseExtensions.swift
+//  BigInt
+//
+//  Created by Koray Koska on 23.06.18.
+//  Copyright © 2018 Boilertalk. All rights reserved.
+//
+
+#if canImport(PromiseKit)
+
+import PromiseKit
+
+#if !Web3CocoaPods
+    import Web3
+#endif
+
+public extension SolidityInvocation {
+
+    public func call(block: EthereumQuantityTag = .latest) -> Promise<[String: Any]> {
+        return Promise { seal in
+            self.call(block: block, completion: seal.resolve)
+        }
+    }
+
+    public func send(nonce: EthereumQuantity? = nil, from: EthereumAddress, value: EthereumQuantity?, gas: EthereumQuantity, gasPrice: EthereumQuantity?) -> Promise<EthereumData> {
+        return Promise { seal in
+            self.send(nonce: nonce, from: from, value: value, gas: gas, gasPrice: gasPrice, completion: seal.resolve)
+        }
+    }
+
+    public func estimateGas(from: EthereumAddress? = nil, gas: EthereumQuantity? = nil, value: EthereumQuantity? = nil) -> Promise<EthereumQuantity> {
+        return Promise { seal in
+            self.estimateGas(from: from, gas: gas, value: value, completion: seal.resolve)
+        }
+    }
+}
+
+public extension SolidityConstructorInvocation {
+
+    public func send(nonce: EthereumQuantity? = nil, from: EthereumAddress, value: EthereumQuantity = 0, gas: EthereumQuantity, gasPrice: EthereumQuantity?) -> Promise<EthereumData> {
+        return Promise { seal in
+            self.send(nonce: nonce, from: from, value: value, gas: gas, gasPrice: gasPrice, completion: seal.resolve)
+        }
+    }
+}
+
+#endif
