+++ Web3/Classes/ContractABI/Contract/ERC20.swift
@@ -7,6 +7,9 @@
 
 import Foundation
 import BigInt
+#if !Web3CocoaPods
+    import Web3
+#endif
 
 /// Base protocol for ERC20
 public protocol ERC20Contract: EthereumContract {
