+++ Web3/Classes/ContractABI/Contract/Eth+Contract.swift
@@ -6,6 +6,9 @@
 //
 
 import Foundation
+#if !Web3CocoaPods
+    import Web3
+#endif
 
 public enum ContractParsingError: Error, LocalizedError {
     case invalidKey
